parse tile default props, fix a warning, fix an infinite loop in tmx parsing

+++ game/scene_game.cpp
@@ -37,7 +37,7 @@ void GameScene::Update(float dt) {
 
 void GameScene::Render() {
 	DC_Clear();
-	DC_SetTransform(inf->width / inf->gameWidth, 0, 0, inf->width / inf->gameWidth, 0, 0, true);
+	DC_SetTransform((float)(inf->width / inf->gameWidth), 0, 0, (float)(inf->width / inf->gameWidth), 0, 0, true);
 
 	DC_DrawMapLayer(0, 0.0f, 0.0f - map->height * map->tile_height + inf->gameHeight);
